+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 2.6)
 
-add_definitions(-std=c++11)
+add_definitions(-std=c++14)
 find_package(Boost REQUIRED COMPONENTS iostreams program_options filesystem)
 find_package(SDL REQUIRED)
 
@@ -16,7 +16,7 @@ target_link_libraries(mdcii-inselbmp boost_system boost_iostreams boost_filesyst
 add_executable(mdcii-weltbmp weltbmp.cpp bildspeicher.cpp bildspeicher_pal8.cpp block.cpp insel.cpp strukturen.cpp welt.cpp bebauung.cpp grafiken.cpp files.cpp version.cpp)
 target_link_libraries(mdcii-weltbmp boost_system boost_iostreams boost_filesystem)
 
-add_executable(mdcii-sdltest sdltest.cpp bildspeicher.cpp bildspeicher_pal8.cpp block.cpp insel.cpp strukturen.cpp welt.cpp kamera.cpp karte.cpp spielbildschirm.cpp bebauung.cpp grafiken.cpp files.cpp version.cpp)
+add_executable(mdcii-sdltest sdltest.cpp bildspeicher.cpp bildspeicher_pal8.cpp block.cpp insel.cpp strukturen.cpp welt.cpp kamera.cpp karte.cpp spielbildschirm.cpp bebauung.cpp grafiken.cpp files.cpp version.cpp cod_parser.cpp)
 target_link_libraries(mdcii-sdltest SDL boost_system boost_iostreams boost_program_options boost_filesystem boost_regex cod_proto protobuf)
 if(WIN32)
   target_link_libraries(mdcii-sdltest mingw32 winmm dxguid SDLmain)
@@ -36,5 +36,5 @@ target_link_libraries(mdcii-zeitext boost_iostreams boost_program_options)
 add_executable(mdcii-codcat codcat.cpp)
 target_link_libraries(mdcii-codcat boost_iostreams)
 
-add_executable(cod_parser_test test/cod_parser_test/cod_parser_test.cpp files.cpp proto/cod.pb.cc)
+add_executable(cod_parser_test test/cod_parser_test/cod_parser_test.cpp files.cpp cod_parser.cpp)
 target_link_libraries(cod_parser_test boost_system boost_iostreams boost_program_options boost_filesystem boost_regex cod_proto protobuf)
